apps/ocr_service/requirements.txt
fastapi
uvicorn[standard]
pytesseract
Pillow

